# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification for Catálogo Publiex Maracaibo
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /products:
    get:
      operationId: listProducts
      tags: [products]
      summary: List all products
      parameters:
        - name: search
          in: query
          required: false
          schema:
            type: string
          description: Search by name
        - name: category
          in: query
          required: false
          schema:
            type: string
          description: Filter by category
        - name: slug
          in: query
          required: false
          schema:
            type: string
          description: Match one exact product slug
      responses:
        "200":
          description: List of products
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Product"
    post:
      operationId: createProduct
      tags: [products]
      summary: Create a product
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProductInput"
      responses:
        "201":
          description: Created product
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
  /product-keywords:
    get:
      operationId: listProductKeywords
      tags: [products]
      summary: List reusable product keywords
      security:
        - cookieAuth: []
      parameters:
        - name: search
          in: query
          required: false
          schema:
            type: string
            maxLength: 60
          description: Filter keywords by text
      responses:
        "200":
          description: Reusable product keywords
          content:
            application/json:
              schema:
                type: array
                items:
                  type: string
                  minLength: 1
                  maxLength: 60
        "401":
          description: Authentication required
    delete:
      operationId: deleteProductKeyword
      tags: [products]
      summary: Remove a reusable product keyword
      security:
        - cookieAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [keyword]
              properties:
                keyword:
                  type: string
                  minLength: 1
                  maxLength: 60
      responses:
        "204":
          description: Keyword removed from all products
        "400":
          description: Invalid keyword
        "401":
          description: Authentication required
  /products/{id}:
    get:
      operationId: getProduct
      tags: [products]
      summary: Get a product by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Product detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
        "404":
          description: Not found
    patch:
      operationId: updateProduct
      tags: [products]
      summary: Update a product
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProductUpdate"
      responses:
        "200":
          description: Updated product
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
        "404":
          description: Not found
    delete:
      operationId: deleteProduct
      tags: [products]
      summary: Delete a product
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
        "404":
          description: Not found
...
components:
  schemas:
...
    ProductSize:
      type: object
      required: [name, price]
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 80
          description: Size name (e.g. "Media carta", "Carta", "Oficio")
        price:
          type: number
          minimum: 0
          maximum: 1000000
          description: Price in USD for this size
        measurements:
          type: string
          minLength: 1
          maxLength: 120
          description: Product measurements in centimeters (e.g. "30 × 40 × 15 cm")
        acrylicThicknessCm:
          type: number
          minimum: 0
          maximum: 100
          multipleOf: 0.01
          description: Acrylic thickness in centimeters
        acrylicColor:
          type: string
          minLength: 1
          maxLength: 60
          description: Acrylic color
    DiscountTier:
      type: object
      required: [threshold, amountOffUsd]
      properties:
        threshold:
          type: integer
          minimum: 1
          description: Minimum quantity for this discount tier
        amountOffUsd:
          type: number
          minimum: 0
          maximum: 1000000
          multipleOf: 0.01
          description: Fixed discount in USD applied to each unit at this level
    Product:
      type: object
      required: [id, name, slug, category, images, keywords, basePrice, deliveryTime, stockQuantity, sizes, createdAt, updatedAt]
      properties:
        id:
          type: integer
        name:
          type: string
        slug:
          type: string
        description:
          type: ["string", "null"]
        keywords:
          type: array
          maxItems: 20
          items:
            type: string
            minLength: 1
            maxLength: 60
          description: Search terms associated with the product
        category:
          type: string
        images:
          type: array
          items:
            type: string
          description: Array of image URLs
        basePrice:
          type: number
          description: Base price in USD (lowest size or default)
        deliveryTime:
          type: string
          description: Estimated delivery time (e.g. "3-5 días hábiles")
        stockQuantity:
          type: integer
          minimum: 0
          description: Units currently available. Zero means made to order.
        sizes:
          type: array
          items:
            $ref: "#/components/schemas/ProductSize"
          description: Available sizes with individual prices
        discountTiers:
          type: array
          items:
            $ref: "#/components/schemas/DiscountTier"
          description: Bulk discount levels ordered by minimum quantity
        discountThreshold:
          type: ["integer", "null"]
          description: Minimum quantity to trigger bulk discount (null = no discount)
        discountPercent:
          type: ["number", "null"]
          description: Legacy percentage discount field; new offers use amountOffUsd
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
...
    ProductInput:
      type: object
      required: [name, slug, category, images, basePrice, deliveryTime, stockQuantity, sizes]
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 120
        slug:
          type: string
          minLength: 1
          maxLength: 160
        description:
          type: string
          maxLength: 4000
        keywords:
          type: array
          maxItems: 20
          items:
            type: string
            minLength: 1
            maxLength: 60
          description: Search terms associated with the product
        category:
          type: string
          minLength: 1
          maxLength: 80
        images:
          type: array
          maxItems: 10
          items:
            type: string
        basePrice:
          type: number
          minimum: 0
          maximum: 1000000
        deliveryTime:
          type: string
          minLength: 1
          maxLength: 120
        stockQuantity:
          type: integer
          minimum: 0
          description: Units currently available. Zero means made to order.
        sizes:
          type: array
          maxItems: 20
          items:
            $ref: "#/components/schemas/ProductSize"
        discountTiers:
          type: array
          items:
            $ref: "#/components/schemas/DiscountTier"
          description: Bulk discount levels ordered by minimum quantity
        discountThreshold:
          type: integer
        discountPercent:
          type: number
          multipleOf: 0.01
    ProductUpdate:
      type: object
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 120
        slug:
          type: string
          minLength: 1
          maxLength: 160
        description:
          type: string
          maxLength: 4000
        keywords:
          type: array
          maxItems: 20
          items:
            type: string
            minLength: 1
            maxLength: 60
          description: Search terms associated with the product
        category:
          type: string
          minLength: 1
          maxLength: 80
        images:
          type: array
          maxItems: 10
          items:
            type: string
        basePrice:
          type: number
          minimum: 0
          maximum: 1000000
        deliveryTime:
          type: string
          minLength: 1
          maxLength: 120
        stockQuantity:
          type: integer
          minimum: 0
          description: Units currently available. Zero means made to order.
        sizes:
          type: array
          maxItems: 20
          items:
            $ref: "#/components/schemas/ProductSize"
        discountTiers:
          type: array
          items:
            $ref: "#/components/schemas/DiscountTier"
          description: Bulk discount levels ordered by minimum quantity
        discountThreshold:
          type: ["integer", "null"]
        discountPercent:
          type: ["number", "null"]
          multipleOf: 0.01
...